+++ algorithmTestingEnvironment/index.html
@@ -8,5 +8,5 @@
 <body>
     
 </body>
-<script  src="./index.js"></script>
+<script  src="./seed.js"></script>
 </html>
