+++ algorithmTestingEnvironment/mongoTesting/config/connection.js
@@ -1,0 +1,7 @@
+const mongoose = require('mongoose');
+
+mongoose.connect(
+  process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/schools-db'
+);
+
+module.exports = mongoose.connection;
