+++ algorithmTestingEnvironment/mongoTesting/models/index.js
@@ -1,0 +1,14 @@
+const User = require('./User');
+const Post = require('./Post');
+const Profile = require('./Profile');
+const Notification = require('./Notification')
+const Comment = require('./Comment')
+
+
+module.exports = {
+    User,
+    Post,
+    Profile,
+    Comment,
+    Notification
+};
