+++ algorithmTestingEnvironment/mongoTesting/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "apollo-school",
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js",
+    "seed": "node seeds/seed.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@apollo/server": "^4.7.1",
+    "express": "^4.17.2",
+    "graphql": "^16.6.0",
+    "mongoose": "^7.0.2"
+  }
+}
