+++ algorithmTestingEnvironment/mongoTesting/schemas/resolvers.js
@@ -1,0 +1,11 @@
+const {User} = require('../models');
+
+const resolvers = {
+  Query: {
+    user: async () => {
+      return await User.find({})
+    }
+  }
+};
+
+module.exports = resolvers;
