+++ client/src/main.jsx
@@ -11,6 +11,8 @@ import NotificationPage from './pages/NotificationPage.jsx';
 import MakeProfile from './pages/TestCreateProfile.jsx';
 
 
+
+
 const router = createBrowserRouter([
   {
     path: '/',
