+++ client/src/pages/NotificationPage.jsx
@@ -4,7 +4,7 @@ import Feed from "../components/PostedContent/Feed";
 import NotifyHeader from "../components/notifications/NotifyHeader";
 import MyButton from "../components/profile/MyButton";
 
-const NotificationPage = () => {
+const NotificationPage = ({}) => {
     //logic to get feed of users profile
     
     // every time this page is loaded it needs a profile to display
@@ -18,10 +18,10 @@ const NotificationPage = () => {
     // nothing will change from user profile to user profile except the data that is passed in as a prop and the follow button will not be present if the user is viewing their own profile
   
     //any changes a user wants to make to their own profile will be handled in the settings page
-  
-  //i need sample data to test this page with
-  // this will be an array of notifications that wil each have a userOfOrigin, textContent, date, ProfilePicture, and unique id
-
+    
+    //i need sample data to test this page with
+    // this will be an array of notifications that wil each have a userOfOrigin, textContent, date, ProfilePicture, and unique id
+console.log(window.user)
     const notificationArray = [
         {
             userOfOrigin: "User",
@@ -60,6 +60,7 @@ const NotificationPage = () => {
         },
         {
             userOfOrigin: "User",
+            userOfOriginId: "1234",
             textContent: "liked your post",
             date: "2021-08-2",
             profilePicture: "https://i.imgur.com/0LKZQYM.jpg",
@@ -74,13 +75,17 @@ const NotificationPage = () => {
         }
     ]
   
+    //query the database for the users profile data using user props 
+    //this will be the users profile data
+
+
     return (
       <div className="bg-bkg-1 text-content">
         <Header />
          <div style={{marginTop:'55px'}}>
             <NotifyHeader />
         </div>
-        <Feed dataArray={notificationArray } type={"notification"} feedToUse={"notification"} /> 
+        <Feed dataArray={notificationArray} type={"notification"} feedToUse={"notification"} /> 
          <div className="flex-1 p-0 pb-4">
         <MyButton type={"bordered"} content={"dismiss all"} shape={"square"} size={"full"} action={null} />
         </div>
@@ -91,4 +96,5 @@ const NotificationPage = () => {
   
   export default NotificationPage
 ;
-  
+  //loggedInUser={loggedInUser}
+  // i need to pass in as props throughout application
